extract itemToCsvRow & csvRowToItem
